# models/error_response.go
package models

// ErrorResponse represents the structure of an error response
type ErrorResponse struct {
	Error string `json:"error"`
}
type MessageResponse struct {
	Message string `json:"message"`
}